Added 3 failing tests: always.finally, never.finally, and thereis.finally
Ignore-this: f3e9d3983567eb04ce2429b310bee837

I'm not sure about what is their contract, but it was surprising to me
that the FINALLY block is not executed.

Arguably if you're expected to avoid sideffects in a FINALLY block,
then you cannot/shouldn't change the return value, so the entire
FINALLY block can be skipped then.

Either way, this behavior should be documented with the tests, so
if the current behavior is expected, then change the tests
accordingly (as opposed to deleting them).

darcs-hash:c0b172b51bc40943946c86d8ab76297807d26594