Emit deprecation warning on use of "count."
The use of count, which hijacks a function definition in the CL package,
violates a design principle articulated in the manual and paper.
Instead, only "counting" should be supported.

When a program uses COUNT in the context of ITERATE, this will emit a
deprecation warning. The intent is to remove support for COUNT as a
synonym for COUNTING in the next version of ITERATE.